Make exit code more useful
We were previously using the number of failures and errors as the exit
code.  Since exit codes are 8-bit integers, this can result in a false
success if there are an exact multiple of 256 failures and errors.

Now, we just exit with a code of 1 if there are problems, and 0
otherwise.